requirements.txt
allure-pytest
assertpy
pybatfish
pydantic>=2,<3
pytest
pyyaml
